add tools file
